chore: util -> utils 폴더명 변경 및 export 방식 변경
